requirements.txt
flask
cairosvg
torch
transformers
diffusers
Pillow